+++ src/kodo_python/decoder.hpp
@@ -16,6 +16,9 @@
 
 #include <sak/storage.hpp>
 
+#include <string>
+#include <vector>
+
 #include "coder.hpp"
 
 namespace kodo_python
@@ -76,13 +79,6 @@ namespace kodo_python
     }
 
     template<class Decoder>
-    bool has_partial_decoding_tracker(Decoder& decoder)
-    {
-        (void) decoder;
-        return kodo::has_partial_decoding_tracker<Decoder>::value;
-    }
-
-    template<class Decoder>
     bool is_partial_complete(Decoder& decoder)
     {
         return kodo::is_partial_complete(decoder);
@@ -100,11 +96,36 @@ namespace kodo_python
         #endif
     }
 
+    template<bool HAS_PARTIAL_DECODING_TRACKER, class Type>
+    struct is_partial_complete_method
+    {
+        template<class DecoderClass>
+        is_partial_complete_method(DecoderClass& decoder_class)
+        {
+            (void) decoder_class;
+        }
+    };
+
+    template<class Type>
+    struct is_partial_complete_method<true, Type>
+    {
+        template<class DecoderClass>
+        is_partial_complete_method(DecoderClass& decoder_class)
+        {
+            decoder_class
+            .def("is_partial_complete", &is_partial_complete<Type>,
+                "Returns true if the decoding matrix should be partially "
+                "decoded.\n\n"
+                "\t:returns: True if the decoding matrix should be partially "
+                "decoded.\n");
+        }
+    };
+
     template<template<class, class> class Coder, class Type>
     struct extra_decoder_methods
     {
         template<class DecoderClass>
-        void operator()(DecoderClass& decoder_class)
+        extra_decoder_methods(DecoderClass& decoder_class)
         {
             (void) decoder_class;
         }
@@ -114,7 +135,7 @@ namespace kodo_python
     struct extra_decoder_methods<kodo::sliding_window_decoder, Type>
     {
         template<class DecoderClass>
-        void operator()(DecoderClass& decoder_class)
+        extra_decoder_methods(DecoderClass& decoder_class)
         {
             decoder_class
             .def("feedback_size", &Type::feedback_size,
@@ -130,7 +151,8 @@ namespace kodo_python
     template<template<class, class> class Coder, class Field, class TraceTag>
     void decoder(const std::string& stack, const std::string& field, bool trace)
     {
-        using namespace boost::python;
+        using boost::python::register_ptr_to_python;
+        using boost::python::arg;
 
         std::string s = "_";
         std::string kind = "decoder";
@@ -138,7 +160,7 @@ namespace kodo_python
         std::string name = stack + s + kind + s + field + trace_string;
 
         typedef Coder<Field, TraceTag> decoder_type;
-        auto decoder_class = coder<Coder,Field,TraceTag>(name)
+        auto decoder_class = coder<Coder, Field, TraceTag>(name)
         .def("recode", &recode<decoder_type>,
             "Recodes a symbol. This function is special for network codes.\n\n"
             "\t:returns: The recoded symbol.\n"
@@ -172,26 +194,19 @@ namespace kodo_python
             "Returns whether the symbol is uncoded or not.\n\n"
             "\t:param index: Index of the symbol to check.\n"
             "\t:returns: True if the symbol is uncoded, and otherwise false.\n"
-        )
-        .def("has_partial_decoding_tracker",
-            &has_partial_decoding_tracker<decoder_type>,
-            "Return whether the decoder contains a partial decoding tracker\n\n"
-            "\t:returns: True if the partial decoding tracker is available, "
-            "and otherwise false.\n"
-        )
-        .def("is_partial_complete", &is_partial_complete<decoder_type>,
-            "Returns true if the decoding matrix should be partially "
-            "decoded.\n\n"
-            "\t:returns: True if the decoding matrix should be partially "
-            "decoded.\n"
         );
 
         void (decoder_type::*decode_symbol2)(uint8_t*, uint32_t) =
             &decoder_type::decode_symbol;
         decoder_class.def("decode_symbol_at_index", decode_symbol2);
 
-        extra_decoder_methods<Coder, decoder_type> extra;
-        extra(decoder_class);
+        is_partial_complete_method<
+            kodo::has_partial_decoding_tracker<decoder_type>::value,
+            decoder_type>
+            is_partial_complete_method(decoder_class);
+
+        extra_decoder_methods<Coder, decoder_type> extra_decoder_methods(
+            decoder_class);
 
         register_ptr_to_python<boost::shared_ptr<decoder_type>>();
     }